Allow transferring a vector to a given memory pool
Summary:
This diff adds an API `BaseVector::transferOrCopyTo(MemoryPool* pool)` to allow
transferring the ownership of a vector to a target memory pool. If a buffer in this
vector cannot be transferred (e.g., BufferView), it is copied to the target pool. After
the call, the vector is entirely owned by the new pool such that the original pool can
be destructed even if this vector is not released yet. (Note that this API transfers a
buffer to the target pool even if it's multiply referenced.)

Differential Revision: D75116719